Transaction class update
Now uses chatId (type: long) instead userId (type: int)

+++ dolgobot/src/com/redeyesgang/DB/Transaction.java
@@ -5,16 +5,16 @@ package com.redeyesgang.DB;
  */
 
 public class Transaction {
-    private int fromId;
-    private int toId;
+    private long fromId;
+    private long toId;
     private int amount;
     private int transactID=-1;
     private String description=null;
 
-    public int getFromId () {
+    public long getFromId () {
         return fromId;
     }
-    public int getToId () {
+    public long getToId () {
         return toId;
     }
     public int getAmount () {
@@ -23,11 +23,11 @@ public class Transaction {
     public String getDescription () {
         return description;
     }
-    public Transaction setFromId (int fromId) {
+    public Transaction setFromId (long fromId) {
         this.fromId = fromId;
         return this;
     }
-    public Transaction setToId (int toId) {
+    public Transaction setToId (long toId) {
         this.toId = toId;
         return this;
     }
@@ -36,7 +36,7 @@ public class Transaction {
         return this;
     }
 
-    public Transaction (int fromId, int toId, int amount) throws TransactionException {
+    public Transaction (long fromId, long toId, int amount) throws TransactionException {
         if (fromId == toId) {
             throw new TransactionException ("Parameters fromId and toId couldn't be equals.");
         } else {
